backend/pyproject.toml
[project]
name = "strathwell-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "bcrypt==4.3.0",
    "cloudinary>=1.44.1",
    "email-validator>=2.3.0",
    "fastapi>=0.128.2",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pgvector>=0.4.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "stripe>=14.3.0",
    "uvicorn>=0.40.0",
]
